wip2

+++ src/kernel/utils/quantization.rs
@@ -97,8 +97,8 @@ pub fn quantize(scale: f32, zero_point: i32, f: f32) -> Result<i32> {
     Ok(q)
 }
 
-pub fn dequantize(scale: f32, zero_point: i32, q: i8) -> Result<f32> {
-    let f = scale * (q as i32 - zero_point) as f32;
+pub fn dequantize(scale: f32, zero_point: i32, q: i32) -> Result<f32> {
+    let f = scale * (q - zero_point) as f32;
     Ok(f)
 }
